リファクタリング 行のすべての処理が終わるまで次の行に進まないよう修正 会話文のない行は自動で次に進むよう修正
