+++ Assets/Novel/Scripts/BackgroundOperation.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 696d567efac603945bdab9733b070ba5
+guid: 274d6ee0adffe2d44ae696875593cc62
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
