+++ Assets/Novel/Scripts/CharacterOperation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d0b209cb063f6b64792230852913d333
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
