+++ Assets/Novel/Scripts/DialogueOperation.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 69c55fe7163b2c349901a2087f3e7e8c
+guid: 9ea3177beb86241498c0ac7461f11db7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
