--- Assets/Novel/Scripts/OperationType.cs
@@ -1,11 +1,0 @@
-﻿namespace Novel
-{
-    public enum OperationType
-    {
-        Dialogue,
-        CharacterLayout,
-        Background,
-        Bgm,
-        Other
-    }
-}
