+++ Assets/Novel/Scripts/OtherOperation.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Novel
+{
+    public class OtherOperation : MonoBehaviour
+    {
+        public void ExecuteOtherOperation(OtherData otherData)
+        {
+
+        }
+    }
+}
