+++ Assets/Novel/Scripts/OtherOperation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1680801e8afe8bd4ebfbc2aca340163a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
