+++ Assets/Novel/Scripts/SoundOperation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9332f2ff4e7d25d45a090d9dc35fcbd1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
